Implemented OpenAI integration and data persistence
This commit upgrades the 'BestReads Recommendations' project with new capabilities. It integrates OpenAI API into OpenAIClientService to get book embeddings for recommendations. Also, EntityFramework Core is added for data persistence, along with base repository for generic database operations. The connection string for database is stored in appsettings. The Book class was moved to Model folder and updated with new properties, and IEntityWithId interface was added. New IndustryIdentifier and IRepository classes were introduced. Also, Functionality was added in BookEmbeddingService to call GetEmbeddingsAsync from OpenAiClient and return the values.