Add Exception for unsuccessful OpenAI Embedding Request
Introduced an exception `EmbeddingRequestException` that would be thrown when the OpenAI client fails to return any data. This change is to make the service more robust and easier to debug. Also, the `IOpenAICleint` interface and its implementation have been adjusted to return a list of floats directly, simplifying the interface and common usage. Test coverage has been expanded to include these new changes.