Add BookRecommendationServiceTests and refine usings
This commit includes the creation of BookRecommendationServiceTests, which will comprehensively test all functionalities of the BookRecommendationService. To keep the code clean and to follow best practices, unused 'using' statements were removed from the BookEmbeddingServiceTests file. Furthermore, 'Moq' and 'FluentAssertions' were added to GlobalUsings for them to be available for all test files. 'Moq' was also removed from BookRecommendationsControllerTests because it's now globally available.