Add docker publishing to GitHub CI workflow
The CI workflow was updated to include steps to publish the application to Docker Hub. This change is made to automate the process of building and pushing a docker image whenever there is a push to the master branch. This will ensure that the most recent version of the app is always available on Docker Hub.