Add GitHub action for updating Azure App Service
This commit adds a new job in the GitHub workflow contained in CI.yml. The job, 'update-azure-app-service', is triggered on pushes to master. It logs into Azure using stored secrets and then restarts an Azure App service. This change aims to automate the update process of the Azure App Service whenever code on the master branch is changed.