Refactor test suite to use Results over ServiceResponses
In the test classes, the use of ServiceResponse class has been replaced with Result class in order to align the testing structure with application changes. This includes the mocks and assertions in test scenarios. Also, the identifier type "ASIN" is removed in faked Book objects to match with the domain constraints. This ensures the test data generated is more realistic.